Create a directory for acceptance test
