+++ build.gradle
@@ -42,6 +42,8 @@ dependencies {
     testImplementation "com.android.tools.build:gradle:$agpVersion"
 }
 
+// For testing. The classpaths of this plugin must be specified
+
 task createClasspathManifest {
     def outputDir = file("$buildDir/$name")
 
@@ -59,6 +61,8 @@ dependencies {
 }
 
 test.dependsOn(['createClasspathManifest'])
+
+sourceSets.test.java.srcDirs += ["src/test/acceptance"]
 
 test {
     testLogging {
